Remove test inhereitance
